Extract chat panel helpers
