+++ tests/test-correctness-phase2.js
@@ -129,6 +129,7 @@ const pwaAppShellAssets = [
   '/js/chat-threads.js',
   '/js/chat-attestation.js',
   '/js/chat-actions.js',
+  '/js/chat-panel.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
